+++ micom/community.py
@@ -230,7 +230,7 @@ class Community(cobra.Model):
                 % (100.0 * self.__db_metrics["found_abundance_fraction"])
             )
             if self.__db_metrics["found_abundance_fraction"] < 0.5:
-                logger.warning(
+                logger.info(
                     "Less than 50%% of the abundance could be matched to the "
                     "model database. Model `%s` may not be representative "
                     "of the sample" % self.id