+++ micom/viz/interactions.py
@@ -3,11 +3,14 @@
 from datetime import datetime
 import json
 from micom.interaction import interactions, summarize_interactions, MES
-from micom.logger import logger
 from micom.workflows import GrowthResults
 from micom.viz.core import Visualization
 import pandas as pd
 import re
+
+import logging
+
+logger = logging.getLogger(__name__)
 
 UNITS = {
     "flux": "mmol/[gDW·h]",
@@ -85,7 +88,7 @@ def plot_mes(
     number of producers and consumers for a given metabolite in a sample.
 
     $$
-    MES = 2\cdot\frac{|p||c|}{|p| + |c|}
+    MES = 2\\cdot\\frac{|p||c|}{|p| + |c|}
     $$
 
     Parameters