+++ micom/workflows/build.py
@@ -26,6 +26,7 @@ def _reduce_group(df):
 def build_and_save(args):
     """Build a single community model."""
     s, tax, db, out, cutoff, solver = args
+
     if os.path.exists(out):
         com = load_pickle(out)
     else:
@@ -102,6 +103,22 @@ def build(
     else:
         os.makedirs(out_folder)
 
+    # Several checks if the taxonomy table makes sense
+    sample_abundances = taxonomy.groupby("sample_id").abundance.sum()
+    if any(sample_abundances == 0):
+        bad = sample_abundances.index[sample_abundances == 0]
+        logger.warning(
+            "The following samples sum to a zero abundance and will be excluded: "
+            f"{', '.join(bad)}"
+        )
+        taxonomy = taxonomy[~taxonomy.sample_id.isin(bad)]
+    if "file" in taxonomy.columns and model_db is not None:
+        logger.warning(
+            "The table includes a `file` column even though a model database "
+            "is used. Will ignore it and use the model database instead."
+        )
+        del taxonomy["file"]
+
     samples = taxonomy.sample_id.unique()
     out_path = pd.Series({s: os.path.join(out_folder, s + ".pickle") for s in samples})
     args = [
@@ -110,17 +127,37 @@ def build(
     ]
     res = workflow(build_and_save, args, threads)
     metrics = pd.concat(res)
-    taxonomy = (
+    manifest = (
         taxonomy.groupby("sample_id")
         .apply(_reduce_group)
         .dropna(axis=1)
         .reset_index(drop=True)
     )
-    taxonomy = taxonomy.loc[:, ~taxonomy.columns.isin(RANKS)]
-    taxonomy["file"] = taxonomy.sample_id + ".pickle"
-    taxonomy = pd.merge(taxonomy, metrics, on="sample_id")
-    taxonomy.to_csv(os.path.join(out_folder, "manifest.csv"), index=False)
-    return taxonomy
+    manifest = manifest.loc[:, ~manifest.columns.isin(RANKS)]
+    manifest["file"] = manifest.sample_id + ".pickle"
+    manifest = pd.merge(manifest, metrics, on="sample_id")
+
+    if model_db is not None:
+        if any(manifest.found_taxa == 0):
+            missing = manifest.sample_id[manifest.found_taxa == 0]
+            logger.warning(
+                "The following samples had no taxon matches in the model "
+                "database and will be excluded. "
+                "We recommend to verify that the taxon names and "
+                "ranks match the database and version. "
+                f"Missing samples: {', '.join(missing)} ."
+            )
+            manifest = manifest[manifest.found_taxa > 0]
+        frac = manifest.found_abundance_fraction
+        if any((frac > 0) & (frac < 0.5)):
+            low = manifest.sample_id[(frac > 0) & (frac < 0.5)]
+            logger.warning(
+                "Less than 50%% of the abundance could be matched to the "
+                f"model database for these samples: {', '.join(low)} ."
+            )
+
+    manifest.to_csv(os.path.join(out_folder, "manifest.csv"), index=False)
+    return manifest
 
 
 REQ_FIELDS = pd.Series(