+++ micom/workflows/grow.py
@@ -3,7 +3,6 @@
 from cobra.util.solver import interface_to_str, OptimizationError
 from micom import load_pickle
 from micom.annotation import annotate_metabolites_from_exchanges
-from micom.logger import logger
 from micom.media import minimal_medium
 from micom.workflows.core import workflow
 from micom.workflows.results import GrowthResults
@@ -11,6 +10,9 @@ from micom.workflows.media import process_medium
 from os import path
 import pandas as pd
 import warnings
+import logging
+
+logger = logging.getLogger(__name__)
 
 warnings.filterwarnings("ignore", category=pd.errors.PerformanceWarning)